ml-services/ocr-service/models.py
"""
OCR model implementations for handwriting recognition.

This module contains different OCR model implementations:
1. TrOCR (Microsoft's Transformer-based OCR)
2. EasyOCR (backup/comparison)
3. Tesseract OCR (fallback)
"""

import torch
from transformers import TrOCRProcessor, VisionEncoderDecoderModel
from PIL import Image
import easyocr
import logging

logger = logging.getLogger(__name__)

class TrOCRModel:
    """
    TrOCR implementation for handwriting recognition.
    Uses Microsoft's pre-trained transformer model.
    """
    
    def __init__(self, model_name="microsoft/trocr-base-handwritten"):
        logger.info(f"Loading TrOCR model: {model_name}")
        self.processor = TrOCRProcessor.from_pretrained(model_name)
        self.model = VisionEncoderDecoderModel.from_pretrained(model_name)
        self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
        self.model.to(self.device)
        logger.info(f"TrOCR model loaded on {self.device}")
    
    def extract_text(self, image: Image.Image) -> tuple[str, float]:
        """
        Extract text from an image using TrOCR.
        
        Args:
            image: PIL Image object
        
        Returns:
            tuple: (extracted_text, confidence_score)
        """
        try:
            # Preprocess image
            pixel_values = self.processor(image, return_tensors="pt").pixel_values
            pixel_values = pixel_values.to(self.device)
            
            # Generate text
            generated_ids = self.model.generate(pixel_values)
            generated_text = self.processor.batch_decode(generated_ids, skip_special_tokens=True)[0]
            
            # TODO: Calculate actual confidence score
            confidence = 0.85
            
            return generated_text, confidence
        except Exception as e:
            logger.error(f"TrOCR extraction error: {str(e)}")
            raise

class EasyOCRModel:
    """
    EasyOCR implementation as backup/comparison model.
    """
    
    def __init__(self, languages=['en']):
        logger.info(f"Loading EasyOCR model for languages: {languages}")
        self.reader = easyocr.Reader(languages, gpu=torch.cuda.is_available())
        logger.info("EasyOCR model loaded")
    
    def extract_text(self, image_path: str) -> tuple[str, float]:
        """
        Extract text from an image using EasyOCR.
        
        Args:
            image_path: Path to image file
        
        Returns:
            tuple: (extracted_text, confidence_score)
        """
        try:
            results = self.reader.readtext(image_path)
            
            if not results:
                return "", 0.0
            
            # Combine all detected text
            text = " ".join([item[1] for item in results])
            
            # Average confidence
            confidence = sum([item[2] for item in results]) / len(results)
            
            return text, confidence
        except Exception as e:
            logger.error(f"EasyOCR extraction error: {str(e)}")
            raise

class OCRPreprocessor:
    """
    Image preprocessing utilities for better OCR accuracy.
    """
    
    @staticmethod
    def preprocess_image(image: Image.Image) -> Image.Image:
        """
        Preprocess image for better OCR results.
        
        TODO: Implement preprocessing steps:
        - Denoise
        - Contrast enhancement
        - Binarization
        - Deskew
        """
        return image
